Draw next-move paths of projectiles
Because of the way I'm using Graphics (which...may not be the correct
way? I'm treating it as a single drawer whereas I'm moderately sure
you're supposed to group them together, with each graphics being a
composable section - however, the *example* uses it as a single drawer,
so!?!?) it's not very easy to make incremental removals of paths. So it
draws them all at once.

This may be...not how I want to end up doing it.