Fix hard crash on player projectile
Obviously, 0 messes with the division in the projection. It
looks *really* janky the way I fixed it, but I can clean up the visuals
later.

This guy is clearly sick and he's making horrific hacking noises over
there at the computer and it's really making me a little queasy. Maybe I
should go to the quiet room or something.