Move AI into own NS
Back from the dead. And by dead, I mean an upper respiratory infection,
which isn't actually that high on the scale of things that make you
dead. So, could be worse!

Still it was miserable and I couldn't sleep so that kinda was less than
ideal. Anyways, enough about my terrible health. What were we doing? Oh,
right, AI.

The AI will be composed into the FoeAI component on creation of the
entity.