Add a stopApproachDistance behaviour
This is copying the behaviour in my old 7DRL, but what I'd like to later
do is a "Keep At" distance.

On the other hand if the enemy is just as fast as you it'd be a heck of
a Benny Hill situation and make you hate life...so that's for later.