Fix projectile first turn & drawing bugs
The first turn wasn't properly being taken by projectiles due to a falsy
zero value leading to an incorrect assignment for TTL.

The drawing bug was because the draw happens while the player object is
being removed from the entity when it dies, but as the game has not yet
transitioned. Added a check condition to prevent issues.