Pull most ship attributes into fixed object map
I am unsure of the common usage of map in JS; the Microsoft
documentation implies that maps should be used for homogeneous data
types that are operated over in a for-each fashion, and objects for
cases where you operate differently based on key. Well, this is
certainly the second case, but if this were Clojure the 'keys have
semantics' is a perfectly valid and encouraged usage.

Community norms! I should find somebody who knows JS to help me out with
this stuff at some point. Maybe I can put it up in Rogue Scholars.