Add reverser path
Should probably have tests. Especially since this is a translation of
some code which...I don't know if it technically worked correctly in the
Python version!

On the other hand, that is the version I put out publicly so technically
if I want to mimic the gameplay of the previous version, I should be
keeping any bugs.