Fix position tagging: hide x, y, expose fixed obj
This is...kinda messed up. Okay, here's what was happening.

The set javascript provides compares using ===. Fine. However, arrays of
the same semantic value don't compare well with ===, because it's not a
values-first immutable system like Clojure, so my set tagging removal
wasn't working.

To fix this I added a getCurrentPosition() which will return an
immutable, value-comparable array, and so...that's not very elegant, I
admit. I don't really know the best way to do it, aside from maybe
stringifying all the tags, but I'm not very enthusiastic about it.